refactor: use a more organized file structure
