+++ src/helpers/log.ts
@@ -1,4 +1,4 @@
-import { colors } from './std.ts'
+import { colors } from '../../std.ts'
 
 export enum Level {
     Debug = 0,
