+++ src/project.ts
@@ -3,19 +3,19 @@ import marked from 'https://esm.sh/marked@1.2.0'
 import postcss, { AcceptedPlugin } from 'https://esm.sh/postcss@8.1.4'
 import { minify } from 'https://esm.sh/terser@5.3.2'
 import { safeLoadFront } from 'https://esm.sh/yaml-front-matter@4.1.0'
-import { Request } from './api.ts'
-import { EventEmitter } from './events.ts'
-import { ensureTextFile, existsDirSync, existsFileSync } from './fs.ts'
-import { createHtml } from './html.ts'
-import log from './log.ts'
-import { getPagePath, RouteModule, Routing } from './routing.ts'
-import { colors, ensureDir, fromStreamReader, path, ServerRequest, Sha1, walk } from './std.ts'
-import { compile } from './tsc/compile.ts'
-import type { APIHandler, Config, RouterURL } from './types.ts'
-import util, { hashShort, MB, reHashJs, reHttp, reLocaleID, reMDExt, reModuleExt, reStyleModuleExt } from './util.ts'
-import { createHTMLDocument } from './vendor/deno-dom/document.ts'
-import less from './vendor/less/less.js'
-import { VERSION } from './version.ts'
+import { Request } from '../api.ts'
+import { EventEmitter } from '../events.ts'
+import { ensureTextFile, existsDirSync, existsFileSync } from '../fs.ts'
+import { getPagePath, RouteModule, Routing } from '../routing.ts'
+import { colors, ensureDir, fromStreamReader, path, ServerRequest, Sha1, walk } from '../std.ts'
+import { compile } from '../tsc/compile.ts'
+import type { APIHandler, Config, RouterURL } from '../types.ts'
+import { createHTMLDocument } from '../vendor/deno-dom/document.ts'
+import less from '../vendor/less/less.js'
+import { VERSION } from '../version.ts'
+import { createHtml } from './helpers/html.ts'
+import log from './helpers/log.ts'
+import util, { hashShort, MB, reHashJs, reHttp, reLocaleID, reMDExt, reModuleExt, reStyleModuleExt } from './helpers/util.ts'
 
 interface Module {
     id: string
