+++ src/project_test.ts
@@ -1,6 +1,6 @@
 import { assertEquals } from 'https://deno.land/std/testing/asserts.ts'
+import { path } from '../std.ts'
 import { Project } from './project.ts'
-import { path } from './std.ts'
 
 Deno.test({
     name: 'project build(hello world)',
