+++ src/server/hmr.ts
@@ -1,6 +1,6 @@
 import runtime from 'https://esm.sh/react-refresh@0.8.3/runtime'
-import events from './events.ts'
-import util, { hashShort } from './util.ts'
+import events from '../../events.ts'
+import util, { hashShort } from '../helpers/util.ts'
 
 interface Callback {
     (...args: any[]): void
