+++ src/server/server.ts
@@ -1,11 +1,11 @@
-import { Request } from './api.ts'
-import { existsFileSync } from './fs.ts'
-import { createHtml } from './html.ts'
-import log from './log.ts'
-import { getContentType } from './mime.ts'
-import { injectHmr, Project } from './project.ts'
-import { path, serve, ws } from './std.ts'
-import util, { hashShort } from './util.ts'
+import { path, serve, ws } from '../..//std.ts'
+import { Request } from '../../api.ts'
+import { existsFileSync } from '../../fs.ts'
+import { createHtml } from '../../html.ts'
+import { getContentType } from '../../mime.ts'
+import log from '../helpers/log.ts'
+import util, { hashShort } from '../helpers/util.ts'
+import { injectHmr, Project } from '../project.ts'
 
 export async function start(appDir: string, hostname: string, port: number, isDev = false, reload = false) {
     const project = new Project(appDir, isDev ? 'development' : 'production', reload)
